+++ core/src/test/scala/pureconfig/ConfigConvertChecks.scala
@@ -1,12 +1,15 @@
 package pureconfig
 
+import scala.reflect.ClassTag
+import scala.reflect.runtime.universe._
+
 import com.typesafe.config.{ ConfigRenderOptions, ConfigValue, ConfigValueFactory }
 import org.scalacheck.Arbitrary
 import org.scalactic.Equality
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
 import org.scalatest.{ EitherValues, FlatSpec, Matchers }
-import pureconfig.error.ConfigReaderFailure
-import scala.reflect.ClassTag
+
+import pureconfig.error._
 
 /**
  * Add utilities to a scalatest `FlatSpec` to test `ConfigConvert` instances
@@ -23,8 +26,8 @@ trait ConfigConvertChecks { this: FlatSpec with Matchers with GeneratorDrivenPro
    * because, for instance, multiple representation of `t: T` are possible. Use [[checkRead()]] for those
    * representations.
    */
-  def checkArbitrary[T](implicit cc: ConfigConvert[T], arb: Arbitrary[T], tag: ClassTag[T], equality: Equality[T]): Unit =
-    it should s"read an arbitrary ${tag.runtimeClass.getCanonicalName}" in forAll {
+  def checkArbitrary[T](implicit cc: ConfigConvert[T], arb: Arbitrary[T], tpe: TypeTag[T], equality: Equality[T]): Unit =
+    it should s"read an arbitrary ${tpe.tpe}" in forAll {
       (t: T) =>
         val result = cc.from(cc.to(t))
         result.right.value shouldEqual t
@@ -48,8 +51,8 @@ trait ConfigConvertChecks { this: FlatSpec with Matchers with GeneratorDrivenPro
    * @param cw the `ConfigConvert` used to write a value to a `ConfigValue`. This is the dummy instance used to test `cr`
    * @param arb the `Arbitrary` used to generate values to write a `ConfigValue` via `cw`
    */
-  def checkArbitrary2[T1, T2](f: T2 => T1)(implicit cr: ConfigConvert[T1], cw: ConfigConvert[T2], arb: Arbitrary[T2], tag1: ClassTag[T1], tag2: ClassTag[T2], equality: Equality[T1]): Unit =
-    it should s"read a ${tag1.runtimeClass.getSimpleName} from an arbitrary ${tag2.runtimeClass.getSimpleName}" in forAll {
+  def checkArbitrary2[T1, T2](f: T2 => T1)(implicit cr: ConfigConvert[T1], cw: ConfigConvert[T2], arb: Arbitrary[T2], tpe1: TypeTag[T1], tpe2: TypeTag[T2], equality: Equality[T1]): Unit =
+    it should s"read a ${tpe1.tpe} from an arbitrary ${tpe2.tpe}" in forAll {
       (t2: T2) => cr.from(cw.to(t2)).right.value shouldEqual f(t2)
     }
 
@@ -57,43 +60,59 @@ trait ConfigConvertChecks { this: FlatSpec with Matchers with GeneratorDrivenPro
    * For each pair of value of type `T` and `ConfigValue`, check that `ConfigReader[T].from`
    * successfully converts the latter into to former. Useful to test specific values
    */
-  def checkRead[T](valuesToReprs: (T, ConfigValue)*)(implicit cr: ConfigReader[T], tag: ClassTag[T], equality: Equality[T]): Unit =
+  def checkRead[T](valuesToReprs: (T, ConfigValue)*)(implicit cr: ConfigReader[T], tpe: TypeTag[T], equality: Equality[T]): Unit =
     for ((value, repr) <- valuesToReprs) {
-      it should s"read the value $value of type ${tag.runtimeClass.getSimpleName} " +
+      it should s"read the value $value of type ${tpe.tpe} " +
         s"from ${repr.render(ConfigRenderOptions.concise())}" in {
           cr.from(repr).right.value shouldEqual value
         }
     }
 
   /** Similar to [[checkRead()]] but work on ConfigValues of type String */
-  def checkReadString[T](valuesToStr: (T, String)*)(implicit cr: ConfigReader[T], tag: ClassTag[T], equality: Equality[T]): Unit =
-    checkRead[T](valuesToStr.map { case (t, s) => t -> ConfigValueFactory.fromAnyRef(s) }: _*)(cr, tag, equality)
+  def checkReadString[T](valuesToStr: (T, String)*)(implicit cr: ConfigReader[T], tpe: TypeTag[T], equality: Equality[T]): Unit =
+    checkRead[T](valuesToStr.map { case (t, s) => t -> ConfigValueFactory.fromAnyRef(s) }: _*)(cr, tpe, equality)
 
   /**
    * For each pair of value of type `T` and `ConfigValue`, check that `ConfigWriter[T].to`
    * successfully converts the former into the latter. Useful to test specific values
    */
-  def checkWrite[T](valuesToReprs: (T, ConfigValue)*)(implicit cw: ConfigWriter[T], tag: ClassTag[T], equality: Equality[T]): Unit =
+  def checkWrite[T](valuesToReprs: (T, ConfigValue)*)(implicit cw: ConfigWriter[T], tpe: TypeTag[T], equality: Equality[T]): Unit =
     for ((value, repr) <- valuesToReprs) {
-      it should s"write the value $value of type ${tag.runtimeClass.getSimpleName} " +
+      it should s"write the value $value of type ${tpe.tpe} " +
         s"to ${repr.render(ConfigRenderOptions.concise())}" in {
           cw.to(value) shouldEqual repr
         }
     }
 
   /**
-   * Check that `cc` returns error of type `E` wwhen trying to read each value passed with `values`
+   * Check that `cc` returns error of type `E` when trying to read each value passed with `values`
    *
    * @param values the values that should not be conver
    * @param cr the [[ConfigConvert]] to test
    */
-  def checkFailure[T, E <: ConfigReaderFailure](values: ConfigValue*)(implicit cr: ConfigReader[T], tag: ClassTag[T], eTag: ClassTag[E]): Unit =
+  def checkFailure[T, E <: ConfigReaderFailure](values: ConfigValue*)(implicit cr: ConfigReader[T], tpe: TypeTag[T], eTag: ClassTag[E]): Unit =
     for (value <- values) {
-      it should s"fail when it tries to read a value of type ${tag.runtimeClass.getCanonicalName} " +
+      it should s"fail when it tries to read a value of type ${tpe.tpe} " +
         s"from ${value.render(ConfigRenderOptions.concise())}" in {
           val result = cr.from(value)
           result.left.value.toList should have size 1
           result.left.value.head shouldBe a[E]
+        }
+    }
+
+  /**
+   * For each pair of `ConfigValue` and `ConfigReaderFailures`, check that `cr`
+   * fails with the provided errors when trying to read the provided
+   * `ConfigValue`.
+   */
+  def checkFailures[T](valuesToErrors: (ConfigValue, ConfigReaderFailures)*)(implicit cr: ConfigReader[T], tpe: TypeTag[T]): Unit =
+    for ((value, errors) <- valuesToErrors) {
+      it should s"fail when it tries to read a value of type ${tpe.tpe} " +
+        s"from ${value.render(ConfigRenderOptions.concise())}" in {
+          val result = cr.from(value)
+          val errorList = errors.toList
+          result.left.value.toList.size shouldEqual errorList.size
+          result.left.value.toList should contain theSameElementsAs errorList
         }
     }
 }